[Feature] #3 TabView, 현재 홈화면은 첫번째 탭, 두번째 화면은 히스토리 탭
